Move header, main and footer components into new layout directory.
