+++ rufus-lang.org/src/App.js
@@ -1,8 +1,8 @@
 import React from 'react';
 
-import Footer from './Footer';
-import Header from './Header';
-import Main from './Main';
+import Footer from './layout/Footer';
+import Header from './layout/Header';
+import Main from './layout/Main';
 
 import './App.css';
